+++ packages/isaacscript-common/src/functions/bosses.ts
@@ -15,8 +15,6 @@ import {
   ALL_BOSSES_SET,
   STAGE_TO_COMBINED_BOSS_SET_MAP,
   STAGE_TO_STAGE_TYPE_TO_BOSS_SET_MAP,
-  STORY_BOSS_ENTITY_TYPES_SET,
-  STORY_BOSS_IDS_SET,
 } from "../sets/bossSets";
 import { REPENTANCE_ONLY_BOSS_IDS_SET } from "../sets/repentanceBossIDsSet";
 import { SIN_ENTITY_TYPES_SET } from "../sets/sinEntityTypesSet";
@@ -249,24 +247,6 @@ function getNumBossSegments(
       return DEFAULT_BOSS_MULTI_SEGMENTS;
     }
   }
-}
-
-/**
- * Helper function to determine if the specified entity type is an end-game story boss, like Isaac,
- * Blue Baby, Mega Satan, The Beast, and so on. This is useful because certain effects should only
- * apply to non-story bosses, like Vanishing Twin.
- */
-export function isStoryBoss(entityType: EntityType): boolean {
-  return STORY_BOSS_ENTITY_TYPES_SET.has(entityType);
-}
-
-/**
- * Helper function to determine if the specified boss ID is an end-game story boss, like Isaac, Blue
- * Baby, Mega Satan, The Beast, and so on. This is useful because certain effects should only apply
- * to non-story bosses, like Vanishing Twin.
- */
-export function isStoryBossID(bossID: BossID): boolean {
-  return STORY_BOSS_IDS_SET.has(bossID);
 }
 
 /**